Added new nav to all HTML, updated CSS, JS
New Nav was added to all HTML pages, and CSS was adjusted to accommodate. Menu items were re-coloured to help users with navigation - The current page icon will be highlighted to stand out more than the blue hue used before.

JavaScript function to move between pages was rewritten to use the element's ID as an input, and then pull that value into the link to navigate to the correct page.
Additional JS function was used to toggle the dropdown menu, which is only used on mobile.